added fp consumption to spell item

+++ src/main/java/com/skullmangames/darksouls/common/capability/entity/PlayerCap.java
@@ -126,7 +126,7 @@ public abstract class PlayerCap<T extends Player> extends LivingCap<T>
 		this.fp = value;
 	}
 	
-	public void raiseFP(int value)
+	public void raiseFP(float value)
 	{
 		this.setFP(this.fp + value);
 	}
